enhancement: set no-undefined on libisis.so
shared libraries should not have undefined symbols. this ensures that
every library that libisis needs to be linked against is listed in the ELF
header. this makes it so errors in linking isis, or missing libraries, are
found when isis is built, rather than when someone tries to use libisis.so